FLAG-78: added unit test classes to separate folders
